kafka-consumer/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
pyarrow==14.0.0
python-json-logger==2.0.7
python-snappy